[ARTS-101] Refactor entities to extend a parent audited entity
