+++ role-service/src/main/java/uk/ac/ox/ndph/mts/role_service/controller/RoleController.java
@@ -8,12 +8,14 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import uk.ac.ox.ndph.mts.role_service.controller.dtos.PermissionDTO;
 import uk.ac.ox.ndph.mts.role_service.model.Role;
-import uk.ac.ox.ndph.mts.role_service.model.RoleDTO;
+import uk.ac.ox.ndph.mts.role_service.controller.dtos.RoleDTO;
 import uk.ac.ox.ndph.mts.role_service.model.RoleRepository;
 import uk.ac.ox.ndph.mts.role_service.service.RoleService;
 
 import javax.validation.Valid;
+import java.util.List;
 import java.util.Optional;
 import java.util.logging.Logger;
 
@@ -59,6 +61,8 @@ public class RoleController {
         Role roleEntity = convertDtoToEntity(roleDto);
         return roleService.saveRole(roleEntity);
     }
+
+
