+++ role-service/src/main/java/uk/ac/ox/ndph/mts/role_service/model/AuditedEntity.java
@@ -9,20 +9,37 @@ import org.springframework.data.annotation.LastModifiedBy;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Id;
+import javax.persistence.*;
 import java.time.LocalDateTime;
 
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Audited
-public class Permission extends AuditedEntity {
-    @Id
+
+@MappedSuperclass
+public abstract class AuditedEntity {
+
+
     @Column
+    @CreatedDate
     @Getter
     @Setter
-    private String id;
+    private LocalDateTime createdDateTime;
+
+    @Column
+    @CreatedBy
+    @Getter
+    @Setter
+    private String createdBy;
+
+    @Column
+    @LastModifiedDate
+    @Getter
+    @Setter
+    private LocalDateTime modifiedDateTime;
+
+    @Column
+    @LastModifiedBy
+    @Getter
+    @Setter
+    private String modifiedBy;
+
 
 }
